backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.0
httpx==0.27.0
anthropic==0.34.2
redis==5.0.1
elasticsearch==8.11.0
deepgram-sdk==3.2.7
baseten==0.8.2
python-multipart==0.0.6
aiofiles==23.2.1
